Move chat to components
